rename: 파일 경로 수정
프로필이미지 컴포넌트는 피드에서 말고도 재활용 되어 폴더를 밖으로 뺐고, 폴더 구조 변경하면서 import 바뀐 것 수정
